perf(react-wildcat): Transpile code over multiple workers
Speed up transpiler by spawning multiple workers across available CPUs.
