+++ packages/react-wildcat/cli/wildcatBabel.js
@@ -5,20 +5,14 @@ const fs = require("fs-extra");
 const cwd = process.cwd();
 const pkg = require("../package.json");
 const path = require("path");
-const pathExists = require("path-exists");
 
 const glob = require("glob");
-const slash = require("slash");
 const commander = require("commander");
-const notifier = require("node-notifier");
 const chokidar = require("chokidar");
 const resolve = require("resolve");
-const Batch = require("batch");
 
 const Logger = require("../src/utils/logger");
 const logger = new Logger("👀");
-
-const wildcatConfig = require("../src/utils/getWildcatConfig")(cwd);
 
 function patterns(val) {
     return val.split(" ");
@@ -58,154 +52,12 @@ try {
 
 const util = babel.util;
 
-let babelOptions = {};
-const babelRcPath = path.join(cwd, ".babelrc");
-
-if (pathExists.sync(babelRcPath)) {
-    babelOptions = JSON.parse(fs.readFileSync(babelRcPath));
-}
-
 if (commander.extensions) {
     commander.extensions = util.arrayify(commander.extensions);
 }
 
-function addSourceMappingUrl(code, loc) {
-    return code + "\n//# sourceMappingURL=" + path.basename(loc);
-}
-
-function log(msg) {
-    if (!commander.quiet) {
-        logger.meta(msg);
-    }
-}
-
-function logError(err) {
-    return `${err.name}: ${err.message}\n${err.codeFrame}`;
-}
-
-function write(src, relative, done) {
-    // remove extension and then append back on .js
-    relative = relative.replace(/\.(\w*?)$/, "") + ".js";
-
-    const dest = path.join(commander.outDir, relative);
-
-    const dataOptions = Object.assign({}, babelOptions, {
-        sourceFileName: slash(path.relative(dest + "/..", src)),
-        sourceMapTarget: path.basename(relative)
-    });
-
-    babel.transformFile(src, dataOptions, function transformData(transformErr, data) {
-        if (transformErr) {
-            logger.error(logError(transformErr));
-
-            // Desktop notification
-            notifier.notify({
-                icon: path.resolve(__dirname, "./logo/nfl-engineering-light.png"),
-                title: `${logger.id} ${transformErr.name} in ${pkg.name}`,
-                message: transformErr.message,
-                sound: "Sosumi"
-            });
-
-            return done && done();
-        }
-
-        if (data.ignored) {
-            return done && done();
-        }
-
-        const sourceMaps = babelOptions.sourceMaps || ((babelOptions.env || {}).development || {}).sourceMaps;
-
-        if (data.map && sourceMaps && sourceMaps !== "inline") {
-            const mapLoc = dest + ".map";
-            data.code = addSourceMappingUrl(data.code, mapLoc);
-
-            fs.createOutputStream(mapLoc)
-                .end(JSON.stringify(data.map));
-        }
-
-        fs.createOutputStream(dest)
-            .on("finish", function outputStreamFinish() {
-                log(src + " -> " + dest);
-                return done && done();
-            })
-            .end(data.code);
-    });
-}
-
-function handleFile(src, filename, done) {
-    if (util.shouldIgnore(src, [])) {
-        return done && done();
-    }
-
-    if (util.canCompile(filename, commander.extensions)) {
-        write(src, filename, done);
-    } else if (commander.copyFiles) {
-        let rawDest;
-
-        const dest = path.join(commander.outDir, filename);
-
-        if (commander.binaryToModule) {
-            rawDest = path.join("bin", filename);
-
-            const origin = `${wildcatConfig.generalSettings.staticUrl || ""}/`;
-            const importable = `module.exports = "${origin}${rawDest}";`;
-
-            fs.createOutputStream(dest)
-                .end(importable);
-        }
-
-        fs.createReadStream(src)
-            .pipe(
-                fs.createOutputStream(rawDest || dest)
-                    .on("finish", function outputStreamFinish() {
-                        log(src + " -> " + dest);
-                        return done && done();
-                    })
-            );
-    }
-}
-
-function handle(filename) {
-    pathExists(filename).then(function existsResult(exists) {
-        if (!exists) {
-            return;
-        }
-
-        fs.stat(filename, function statResult(statErr, stats) {
-            if (statErr) {
-                return logger.error(statErr);
-            }
-
-            if (stats.isDirectory(filename)) {
-                const currentDirectory = filename;
-
-                const batch = new Batch();
-                batch.concurrency(20);
-
-                function batchFile(currentFile) {
-                    const src = path.join(currentDirectory, currentFile);
-                    batch.push(function batchJob(done) {
-                        handleFile(src, currentFile, done);
-                    });
-                }
-
-                glob("**", {
-                    cwd: filename,
-                    nodir: true,
-                    ignore: commander.ignore
-                }, function globResult(err, files) {
-                    files.forEach(function individualFile(currentFile) {
-                        batchFile(currentFile);
-                    });
-                    batch.end();
-                });
-            } else {
-                const currentDirectory = filename.replace(`${cwd}/`, "").split("/")[0];
-                write(filename, filename.replace(`${currentDirectory}/`, ""));
-            }
-        });
-    });
-}
+const handleFile = require("./utils/handleFile")(commander);
+const handle = require("./utils/handle")(commander);
 
 if (!commander.watch) {
     let filenames;
